S1-sharp-wallet-copy/requirements.txt
requests>=2.28.0
sqlite3  # Built into Python
web3>=6.0.0
websockets>=11.0.0
python-dotenv>=0.19.0
pandas>=1.5.0
numpy>=1.24.0